pyproject.toml
[project]
name = "hotelbookingsystemfinal"
version = "0.1.0"
description = "Hotel Booking System with FastAPI, SQLModel, Redis, Celery, PostgreSQL"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.116.0",
    "uvicorn[standard]>=0.30.0",
    "sqlmodel>=0.0.24",
    "alembic>=1.13.1", # hash password
    "python-jose[cryptography]>=3.3.0", # JWT
    "pydantic-settings>=2.2.1", # load config từ .env # validate email (nếu có đăng ký user)
    "email-validator>=2.1.0",
    "psycopg[binary]",
    "python-multipart>=0.0.20",
    "bcrypt==4.0.1",
    "passlib[bcrypt]==1.7.4",
    "fastapi-mail>=1.5.0",
    "celery>=5.5.3",
    "redis>=6.4.0",
    "pyotp>=2.9.0",
    "python-dotenv>=1.1.1",
]

# ------------------------------
# Development & Testing Dependencies
# ------------------------------
[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
    "faker>=26.0.0",

    # Code Quality
    "ruff>=0.6.0",
    "black>=24.8.0",
    "mypy>=1.11.0",
    "pre-commit>=3.8.0",

    # Type Hints / Stubs
    "types-python-dateutil>=2.9.0",
    "types-redis>=4.6.0",
]

# ------------------------------
# Ruff (Linter & Formatter)
# ------------------------------
[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

# ------------------------------
# Black Formatter
# ------------------------------
[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'

# ------------------------------
# MyPy (Type Checking)
# ------------------------------
[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

# ------------------------------
# Pytest Configuration
# ------------------------------
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=app",
    "--cov-report=term-missing",
    "--cov-report=html",
]

# ------------------------------
# Dependency Groups for `uv`
# ------------------------------
[dependency-groups]
dev = [
    "black>=25.9.0",
    "faker>=37.12.0",
    "mypy>=1.18.2",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.3",
]
